# fix: Add duration validation and correct sleep range test benchmarks