- Fixed incorrect performance benchmarks in metadata_sleep_range_test.txt
  - Changed min_duration from 19.0 to 26.0 seconds
  - Changed max_duration from 21.0 to 30.0 seconds
  - Benchmarks now correctly reflect sum of sleep durations (1+2.5+5+7.5+10=26s)

- Enhanced intelligent_test_runner.py validation
  - Added min_duration/max_duration validation for ANY test type
  - Previously only validated max_execution_time for performance tests
  - Now properly catches incorrect duration benchmarks
  - Validation runs in Phase 5 before performance-specific checks

This ensures test benchmarks are properly validated and prevents false positives from tests with incorrect timing expectations.

🤖 Generated with [Claude Code](https://claude.com/claude-code)